Extract BaseEntity into its own file
Ticket: AT-6836
