+++ packages/api/orm/entity/ProvisionableEntity.ts
@@ -1,19 +1,6 @@
-import { Column, CreateDateColumn, DeleteDateColumn, PrimaryGeneratedColumn, UpdateDateColumn } from "typeorm";
+import { BaseEntity } from "./BaseEntity";
+import { Column } from "typeorm";
 import { ProvisioningStatus } from "../../models/ProvisioningStatus";
-
-export abstract class BaseEntity {
-  @PrimaryGeneratedColumn("uuid")
-  id: string;
-
-  @CreateDateColumn()
-  createdDate: Date;
-
-  @UpdateDateColumn()
-  updatedDate: Date;
-
-  @DeleteDateColumn()
-  deletedDate: Date;
-}
 
 export abstract class ProvisionableEntity extends BaseEntity {
   @Column({
